Factor out generic extractor out of ISO Camt.053.
Thanks to Rust lang forum, I could make it work without GATs.
https://users.rust-lang.org/t/how-can-i-write-a-generic-matcher/67621/13
